NullLock.acquire should accept wait parameter, return boolean
The interface for acquiring locks specifies that they should accept and
default to wait=True. NullLock wasn't implementing this correctly.
Additionally, the acquire() function needs to return a boolean value
to indicate that the lock was acquired (which in the case of NullLock
is unconditional).

Fixes: #97

Change-Id: I45487cd602b562242e425395728a9ed7e78d797e
Pull-request: https://bitbucket.org/zzzeek/dogpile.cache/pull-requests/50