Do not store empty dict when using Region.get_or_create_multi with should_cache_fn
If `Region.get_or_create_multi` is used with a `should_cache`
function and no value of the current set should be cached,
do not try to store the empty dict in the backend.  It's
unnecessary and also breaks on some backends such as Redis.

Change-Id: Ib7a48d0d7b2f6569e09008f4603f8a355bf9396a
Pull-request: https://bitbucket.org/zzzeek/dogpile.cache/pull-requests/58