Fix __version__ regexp
The quoting style changed after the 0.8.0 release when the 0.8.1
name was added in, in 8476c48eb7b8fd6bf9e83f7c509352c8e326e5cc.
This broke setup.py in master, but fortnuately not in the release
itself (which would be impossible in any case because setup.py
sdist would have failed).

Fixes: #164
Change-Id: Id5aaa994960df68bf5fd415707d3a7546c73faee