Add option for thread local Redis locks
Added option to the Redis backend
:paramref:`.RedisBackend.thread_local_lock`, which when set to False will
disable the use of a threading local  by the ``redis`` module in its
distributed lock service, which is known to interfere with the lock's
behavior when used in an "async" use case, within dogpile this would be
when using the :paramref:`.CacheRegion.async_creation_runner` feature. The
default is conservatively being left at True, but it's likely this should
be set to False in all cases, so a warning is emitted if this flag is not
set to False in conjunction with the distributed lock. Added an optional
argument to :class:`.RedisBackend` that specifies whether or not a
thread-local Redis lock should be used.  This is the default, but it breaks
asynchronous runner compatibility.

Fixes: #171
Closes: #175
Pull-request: https://github.com/sqlalchemy/dogpile.cache/pull/175
Pull-request-sha: 755ae1b7bd826096f9d6df3ba7ab2e20f1ac4ecd

Change-Id: Ica4ed122b86c218e92cdf6174f79ff3870067934