Omit pyproject.toml from installation
Ensured that the "pyproject.toml" file is not included in builds, as the
presence of this file indicates to pip that a pep-517 installation process
should be used.  As this mode of operation appears to be not well supported
by current tools / distros, these problems are avoided within the scope of
dogpile.cache installation by omitting the file.

Fixes: #178
Change-Id: I78bda978a4d2f512ab51f4b4afec3ee3d70adfbf