Allow to configure pymemcache's HashClient retrying
These changes allow us to configure pymemcache's HashClient retrying
mechanisms by exposing his public params. Without these changes that we are
limitated to the default values.

Configuring the internal retrying feature of the HashClient can be
useful in a HA context to properly setup the failover.

The internal retrying feature are a good complement to the retrying
wrapper introduced previously.

[1] https://github.com/pinterest/pymemcache/blob/071191455363943e7d5919701465455e78bb64ae/pymemcache/client/hash.py#L66-L71

Change-Id: Id7aa3d363d281daa640ca34a325454d4c18a8ff0