update pep484 tests
staticmethod can now be type-qualified, so static checkers
are complaining.  however we don't want to do that here
because it's going to fail in Python 3.7.   just use
an ignore type for the one place we do this right now.

Add pep484 tox target as we are getting jenkins to
target this for all projects.

Change-Id: Ic4db0fc7dd0bfe01308027c911f4dcb288bb169e