Moves dead_retry and socket_timeout into the MemcachedBackend class
Moves the ``dead_retry`` argument and the ``socket_timeout`` argument which
were erroneously added to the "set_parameters", where they have no effect, to
be part of the Memcached connection arguments.

Indeed, In my previous patch [1] I proposed to add the ``dead_retry`` and
``socket_timeout`` params to the ``MemcacheArgs`` class. I was wrong. My goal
was to pass these parameters to the client during its initialization to set
the memcached client dead_retry and socket_timeout arguments [2].

By using the MemcacheArgs they are passed to the method calls which is not what
it was requested in the feature request [3]. I misunderstood the goal of
this class (MemcacheArgs).

The ``MemcacheArgs`` class is only inherited by the ``MemcachedBackend``
class and the ``PylibmcBackend`` class. Both libraries doesn't support
``dead_retry`` and ``socket_timeout`` in their methods related to the
memcache API commands (set, get, set_multi, etc), so for this reason I
think we can move those parameters safely.

My previous patch led to issues [4][5] that I'm able to reproduce locally by
using oslo.cache's functional test. These changes fix these issues.

[1] https://github.com/sqlalchemy/dogpile.cache/commit/1de93aab14c1274f20c1f44f8adff3b143c864f6
[2] https://github.com/linsomniac/python-memcached/blob/7942465eba2009927e5d14b4b6dbd48b75780d80/memcache.py#L165
[3] https://github.com/sqlalchemy/dogpile.cache/issues/223
[4] https://bugzilla.redhat.com/show_bug.cgi?id=2103117
[5] https://review.opendev.org/c/openstack/requirements/+/848827