refresh / python 3.8 / 1.3.0
do a full bump of everything, flake8, black etc.
and get minimum version on python 3.8.  pep-621
also enabled as this is very standard now.  Also pifpaf
requires pkg_resources which was removed from python 3.12
so do some rearrangement to get that in the dependencies.
Fix really old and sort of pointless(?) test that loads the
"mako" entrypoint to use importlib.metadata (standard
as of Python 3.8, so we're good).

Apply a hardening of the distributed lock for memcached,
using a mutex when calling client.add() so that multiple
threads calling on this python method are also synchronized

Change-Id: Ic3986070e21c84cdea12d8cfa01cc353f21adf0f