redis: Add username
Redis introduced ACL feature in 4.0.0, and this feature is supported
by redis-py since 3.4.0[1]. When ACL is enabled, authentication
requires username in addition to password.

This adds the username argument to Redis backend and Redis Sentinel
backend, so that username can be passed in arguments along with
password, instead of using client_kwargs and sentinel_kwargs.

[1] https://github.com/redis/redis-py/commit/8df8cd54d135380ad8b3b8807a67a3e6915b0b49