Accept Iterable for backends
A Sequence requires the ability to index (__getitem__) and retrieve the
length (__len__), amongst other options. An Iterable simply required the
ability to iterate (__iter__). This allows us to pass, amongst other
things, dict_keys, to the 'get_multi', 'get_serialized_multi', and
'delete_multi' methods, which seems like a rational thing to want and be
able to do.

Signed-off-by: Stephen Finucane <stephen@that.guru>